requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
google-generativeai==0.3.2
faiss-cpu==1.7.4
numpy==1.24.3
tqdm==4.66.1
networkx==3.2.1
community==1.0.0b1
scikit-learn==1.3.2
pandas==2.1.4
python-multipart==0.0.6